third_party/blink/web_tests/external/wpt/selection/textcontrols/selectionchange-expected.txt

This is a testharness.js-based test.
Found 44 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Modifying selectionStart value of the input element
  assert_equals: expected 1 but got 0
[FAIL] Modifying selectionEnd value of the input element
  assert_equals: expected 1 but got 0
[FAIL] Calling setSelectionRange() on the input element
  assert_equals: expected 1 but got 0
[FAIL] Calling select() on the input element
  assert_equals: expected 1 but got 0
[FAIL] Calling setRangeText() on the input element
  assert_equals: expected 1 but got 0
[FAIL] Setting the same selectionStart value twice on the input element
  assert_equals: expected 1 but got 0
[FAIL] Setting the same selectionEnd value twice on the input element
  assert_equals: expected 1 but got 0
[FAIL] Setting the same selection range twice on the input element
  assert_equals: expected 1 but got 0
[FAIL] Calling select() twice on the input element
  assert_equals: expected 1 but got 0
[FAIL] Calling setRangeText() after select() on the input element
  assert_equals: expected 2 but got 0
[FAIL] Calling setRangeText() repeatedly on the input element
  assert_equals: expected 4 but got 0
[FAIL] Modifying selectionStart value of the disconnected input element
  assert_equals: expected 1 but got 0
[FAIL] Modifying selectionEnd value of the disconnected input element
  assert_equals: expected 1 but got 0
[FAIL] Calling setSelectionRange() on the disconnected input element
  assert_equals: expected 1 but got 0
[FAIL] Calling select() on the disconnected input element
  assert_equals: expected 1 but got 0
[FAIL] Calling setRangeText() on the disconnected input element
  assert_equals: expected 1 but got 0
[FAIL] Setting the same selectionStart value twice on the disconnected input element
  assert_equals: expected 1 but got 0
[FAIL] Setting the same selectionEnd value twice on the disconnected input element
  assert_equals: expected 1 but got 0
[FAIL] Setting the same selection range twice on the disconnected input element
  assert_equals: expected 1 but got 0
[FAIL] Calling select() twice on the disconnected input element
  assert_equals: expected 1 but got 0
[FAIL] Calling setRangeText() after select() on the disconnected input element
  assert_equals: expected 2 but got 0
[FAIL] Calling setRangeText() repeatedly on the disconnected input element
  assert_equals: expected 4 but got 0
[FAIL] Modifying selectionStart value of the textarea element
  assert_equals: expected 1 but got 0
[FAIL] Modifying selectionEnd value of the textarea element
  assert_equals: expected 1 but got 0
[FAIL] Calling setSelectionRange() on the textarea element
  assert_equals: expected 1 but got 0
[FAIL] Calling select() on the textarea element
  assert_equals: expected 1 but got 0
[FAIL] Calling setRangeText() on the textarea element
  assert_equals: expected 1 but got 0
[FAIL] Setting the same selectionStart value twice on the textarea element
  assert_equals: expected 1 but got 0
[FAIL] Setting the same selectionEnd value twice on the textarea element
  assert_equals: expected 1 but got 0
[FAIL] Setting the same selection range twice on the textarea element
  assert_equals: expected 1 but got 0
[FAIL] Calling select() twice on the textarea element
  assert_equals: expected 1 but got 0
[FAIL] Calling setRangeText() after select() on the textarea element
  assert_equals: expected 2 but got 0
[FAIL] Calling setRangeText() repeatedly on the textarea element
  assert_equals: expected 4 but got 0
[FAIL] Modifying selectionStart value of the disconnected textarea element
  assert_equals: expected 1 but got 0
[FAIL] Modifying selectionEnd value of the disconnected textarea element
  assert_equals: expected 1 but got 0
[FAIL] Calling setSelectionRange() on the disconnected textarea element
  assert_equals: expected 1 but got 0
[FAIL] Calling select() on the disconnected textarea element
  assert_equals: expected 1 but got 0
[FAIL] Calling setRangeText() on the disconnected textarea element
  assert_equals: expected 1 but got 0
[FAIL] Setting the same selectionStart value twice on the disconnected textarea element
  assert_equals: expected 1 but got 0
[FAIL] Setting the same selectionEnd value twice on the disconnected textarea element
  assert_equals: expected 1 but got 0
[FAIL] Setting the same selection range twice on the disconnected textarea element
  assert_equals: expected 1 but got 0
[FAIL] Calling select() twice on the disconnected textarea element
  assert_equals: expected 1 but got 0
[FAIL] Calling setRangeText() after select() on the disconnected textarea element
  assert_equals: expected 2 but got 0
[FAIL] Calling setRangeText() repeatedly on the disconnected textarea element
  assert_equals: expected 4 but got 0
Harness: the test ran to completion.
